Revert "ssh: Combine exit code and main command"
This optimizations occasionally breaks the CGroups API.
With this patch in some commands fails reporting that there are too many
values to unpack. This happens for example when we try to execute a
command such as:

   [ -d /some/folder ] || mkdir /some/folder

If the folder already exists, this command returns these lines

---8<---

0
---8<---

Which seems to breack the optimization.

This reverts commit 0a20cec2d9c913a0fa8efd402bf9f7b308b2956e.