cgroups: add high level API for controller/cgroups tracing
For each CGroup we can already trace which tasks are attached.  This is
an extnsion on top of the existing API which allows to easily trace a
set of specified CGroup and/or Controller.

Such new API makes it easy to trace the current CGroup for each task in
a specified list of CGroup controllers and actual CGroups.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>